Zoltan2: Adding parallel multiway merge sort for local sort portion of MJ Sort.
Set the below flags for compilations:
  -D Trilinos_ENABLE_OpenMP:BOOL=$BUILD_OPENMP \
  -D Trilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=OFF \
  -D Trilinos_ENABLE_TESTS:BOOL=ON \
  -D Trilinos_ENABLE_ALL_PACKAGES:BOOL=OFF \
  -D Trilinos_ENABLE_Tpetra:BOOL=ON \
  -D Trilinos_ENABLE_Teuchos:BOOL=ON \
  -D Trilinos_ENABLE_Zoltan2:BOOL=ON \
  -D Zoltan2_ENABLE_OpenMP:BOOL=ON \
  -D Kokkos_ENABLE_OpenMP:BOOL=$BUILD_OPENMP \
  -D Trilinos_ENABLE_Epetra:BOOL=OFF \
  -D Tpetra_INST_OPENMP:BOOL=$BUILD_OPENMP \
  -D Trilinos_ENABLE_OpenMP:BOOL=ON \

+++ packages/zoltan2/test/helpers/Zoltan2_TestHelpers.hpp
@@ -139,8 +139,8 @@ typedef Tpetra::Map<>::node_type znode_t;
     typedef size_t  zgno_t;
 # else  // !TEST_STK_DATA_TYPES
     typedef double zscalar_t;
-    typedef int zlno_t;
-    typedef int zgno_t;
+    typedef long zlno_t;
+    typedef long zgno_t;
 #   define HAVE_EPETRA_DATA_TYPES
 # endif  // TEST_STK_DATA_TYPES
